Don't rely on GC to free nvlists
In practice I've found programs crashing with SIGBUS in
custom_finalize_nvlist().  The reason for this is not clear.  Requiring
explicit frees works around the issue, though at the burden of the
programmer.

The only explanation I've come up with is nvlist_alloc_custom() being
called to wrap something that has already been wrapped. I may revisit
this again.  For now, take inspiration from uring_stubs.c and rely on
the programmer to clean up rather than the GC.